+++ UnitsNet/CustomCode/Quantities/Angle.extra.cs
@@ -10,7 +10,7 @@ namespace UnitsNet
         /// <summary>Get <see cref="RotationalSpeed"/> from <see cref="Angle"/> delta over time delta.</summary>
         public static RotationalSpeed operator /(Angle angle, TimeSpan timeSpan)
         {
-            return RotationalSpeed.FromRadiansPerSecond(angle.Radians / timeSpan.TotalSeconds);
+            return RotationalSpeed.FromRadiansPerSecond(angle.Radians / (decimal)timeSpan.TotalSeconds);
         }
 
         /// <inheritdoc cref="op_Division(UnitsNet.Angle,System.TimeSpan)" />
