+++ UnitsNet/CustomCode/Quantities/Mass.extra.cs
@@ -21,7 +21,7 @@ namespace UnitsNet
         ///     StonePounds related code makes it easier to work with Stone/Pound combination, which are customarily used in the UK
         ///     to express body weight. For example, someone weighs 11 stone 4 pounds (about 72 kilograms).
         /// </summary>
-        private const double StonesInOnePound = 14.0;
+        private const decimal StonesInOnePound = 14.0m;
 
         /// <summary>
         ///     Converts the mass to a customary stone/pounds combination.
@@ -42,27 +42,27 @@ namespace UnitsNet
         /// <summary>
         ///     Get Mass from combination of stone and pounds.
         /// </summary>
-        public static Mass FromStonePounds(double stone, double pounds)
+        public static Mass FromStonePounds(decimal stone, Decimal pounds)
         {
-            return FromPounds(StonesInOnePound*stone + pounds);
+            return FromPounds(StonesInOnePound * stone + pounds);
         }
 
         /// <summary>Get <see cref="MassFlow"/> from <see cref="Mass"/> divided by <see cref="TimeSpan"/>.</summary>
         public static MassFlow operator /(Mass mass, TimeSpan timeSpan)
         {
-            return MassFlow.FromKilogramsPerSecond(mass.Kilograms/timeSpan.TotalSeconds);
+            return MassFlow.FromKilogramsPerSecond(mass.Kilograms / (decimal)timeSpan.TotalSeconds);
         }
 
         /// <summary>Get <see cref="MassFlow"/> from <see cref="Mass"/> divided by <see cref="Duration"/>.</summary>
         public static MassFlow operator /(Mass mass, Duration duration)
         {
-            return MassFlow.FromKilogramsPerSecond(mass.Kilograms/duration.Seconds);
+            return MassFlow.FromKilogramsPerSecond(mass.Kilograms / duration.Seconds);
         }
 
         /// <summary>Get <see cref="Density"/> from <see cref="MassFlow"/> divided by <see cref="Volume"/>.</summary>
         public static Density operator /(Mass mass, Volume volume)
         {
-            return Density.FromKilogramsPerCubicMeter(mass.Kilograms/volume.CubicMeters);
+            return Density.FromKilogramsPerCubicMeter(mass.Kilograms / volume.CubicMeters);
         }
 
         /// <summary>Get <see cref="Volume"/> from <see cref="Mass"/> divided by <see cref="Density"/>.</summary>
@@ -80,13 +80,13 @@ namespace UnitsNet
         /// <summary>Get <see cref="Force"/> from <see cref="Mass"/> times <see cref="Acceleration"/>.</summary>
         public static Force operator *(Mass mass, Acceleration acceleration)
         {
-            return Force.FromNewtons(mass.Kilograms*acceleration.MetersPerSecondSquared);
+            return Force.FromNewtons(mass.Kilograms * acceleration.MetersPerSecondSquared);
         }
 
         /// <summary>Get <see cref="Force"/> from <see cref="Acceleration"/> times <see cref="Mass"/>.</summary>
         public static Force operator *(Acceleration acceleration, Mass mass)
         {
-            return Force.FromNewtons(mass.Kilograms*acceleration.MetersPerSecondSquared);
+            return Force.FromNewtons(mass.Kilograms * acceleration.MetersPerSecondSquared);
         }
     }
 
@@ -99,7 +99,7 @@ namespace UnitsNet
         /// <summary>
         ///     Construct from stone and pounds.
         /// </summary>
-        public StonePounds(double stone, double pounds)
+        public StonePounds(decimal stone, decimal pounds)
         {
             Stone = stone;
             Pounds = pounds;
@@ -108,12 +108,12 @@ namespace UnitsNet
         /// <summary>
         ///     The stone value it was created with.
         /// </summary>
-        public double Stone { get; }
+        public decimal Stone { get; }
 
         /// <summary>
         ///     The pounds value it was created with.
         /// </summary>
-        public double Pounds { get; }
+        public decimal Pounds { get; }
 
         /// <inheritdoc cref="ToString(IFormatProvider)"/>
         public override string ToString()
